Added new progress bars for cooking and selling.

+++ Assets/Scripts/MinMaxTimer.cs
@@ -19,7 +19,7 @@ public class MinMaxTimer : MarioObject
 
 	protected override void Tick()
 	{
-		_timer -= RealDeltaTime;
+		_timer -= (float)RealDeltaTime;
 		if (_timer > 0)
 			return;
 	}
